backend/services/sympy_service.py
"""
SymPy service for symbolic mathematics preprocessing.
"""
import logging
from typing import Optional

logger = logging.getLogger(__name__)

class SymPyService:
    """Service for symbolic mathematics using SymPy"""

    def __init__(self):
        self.sympy_available = False

        # Try to import SymPy
        try:
            import sympy as sp
            from sympy import latex, simplify, expand, factor, solve, diff, integrate, limit
            self.sp = sp
            self.latex = latex
            self.simplify = simplify
            self.expand = expand
            self.factor = factor
            self.solve = solve
            self.diff = diff
            self.integrate = integrate
            self.limit = limit
            self.sympy_available = True
            print("✅ Successfully imported SymPy")
        except ImportError as e:
            print(f"⚠️ SymPy not available: {e}")
            self.sympy_available = False

    def preprocess_latex(self, latex_content: str) -> str:
        """
        Preprocess LaTeX content using SymPy for symbolic computation.

        This is a basic implementation that can be expanded for more sophisticated
        symbolic manipulations like simplification, expansion, solving, etc.
        """
        if not self.sympy_available:
            return latex_content

        try:
            # For now, just return the original content
            # This can be expanded to:
            # - Parse LaTeX expressions
            # - Simplify complex expressions
            # - Compute derivatives/integrals symbolically
            # - Solve equations
            # - Convert between forms

            # Example future implementation:
            # if 'simplify' in latex_content.lower():
            #     # Parse and simplify expressions
            #     pass
            # elif '\\int' in latex_content:
            #     # Compute symbolic integrals
            #     pass

            return latex_content

        except Exception as e:
            logger.warning(f"SymPy preprocessing failed: {e}")
            return latex_content

    def is_available(self) -> bool:
        """Check if SymPy is available"""
        return self.sympy_available

# Global SymPy service instance
sympy_service = SymPyService()